backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
scikit-learn==1.4.0
pandas==2.2.1
joblib==1.3.2
PyPDF2==3.0.1
python-docx==1.1.0